+++ R/download_functions.R
@@ -2,7 +2,7 @@
 #'
 #' @format Named list of all download functions
 #' @include lexicon_afinn.R lexicon_loughran.R lexicon_bing.R lexicon_nrc.R
-#' @include dataset_sentence_polarity.R
+#' @include dataset_sentence_polarity.R dataset_ag_news.R
 #' @noRd
 "download_functions"
 download_functions <- list(
@@ -10,5 +10,6 @@ download_functions <- list(
   sentence_polarity = download_sentence_polarity,
   loughran = download_loughran,
   bing = download_bing,
-  nrc = download_nrc
+  nrc = download_nrc,
+  ag_news = download_ag_news
 )