+++ NEWS.md
@@ -1,5 +1,6 @@
 # textdata (development version)
 
-* `lexicon_nrc()` has been added to give access to the  NRC Emotion lexicon (@juliasilge, #11).
+* `lexicon_nrc()` has been added to give access to the NRC Emotion lexicon (@juliasilge, #11).
+* `dataset_ag_news()` has be added to give access to AG's News Topic Classification Dataset.
 
 # textdata 0.1.0